Fix current note indicator in <NotesList />
